requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
pillow>=9.0.0
tqdm>=4.60.0
jupyter>=1.0.0
opencv-python>=4.0.0